Add files via upload
Divide the Page into 3 rows. 

Divide the first row into two columns. Insert the logo into the first frame and the website header into the right frame.
Divide the 2nd row into 5 columns and insert the hyperlinks such as Home, Login, Registration, Catalogue, Cart, and Contact Us. (When we click on any hyperlink, corresponding web pages should be loaded into the bottom right frame)
Divide the 3rd row into 2 columns. In the left column insert hyperlinks for the services of the website. (When we click on any hyperlink, corresponding web pages should be loaded into the right frame). In the right column write the description of the webpage.